Checkpoint - modulate but leave current server untouched
PR 192 significantly changes ml server, so leaving duplicate code until that's in
